implement "buffer and flush" in cli-wrapper
Implement MVP algo for handling command output from both long and short
running apps. Output from commands is held in a line-oriented buffer and
flushed back to chat after 10 lines of output has accumulated or 2
seconds has passed. This allows for a smooth flow of replies back to the
caller. Hopefully it is suitable for many types of commands, both long
and short running with little or a lot of output =)