Rework registration, dump buffer config from client capabilities
After long days mulling and discussing. This setup should allow the
simplest setup for clients. They don't force nor do they get to choose
on buffering. Instead we have a const (TODO: config) setting that sets
the command backlog for any given redshirt. This simplifies the client
proto and also the server code when it comes to changing the
configuration. Now the last client to register wins, always. But we
won't orphan old clients. Simply change the capability to the latest.
This should be the best compromize for simplicity.